requirements.txt
wheel>=0.22
nameko>=2.9.0
coverage==4.5.1
pytest-cov==2.5.1
codecov==2.0.15
